validations folder refactor
