+++ src/app/(protected)/[group]/[subGroup]/[instance]/(admin-panel)/(stage-4)/algorithm-details/details-table.tsx
@@ -6,7 +6,7 @@ import {
   TableHeader,
   TableRow,
 } from "@/components/ui/table";
-import { ServerResponse } from "@/lib/validations/algorithm";
+import { ServerResponse } from "@/lib/validations/matching";
 
 export function DetailsTable({ data }: { data: ServerResponse }) {
   return (
