+++ src/app/help/page.tsx
@@ -1,4 +1,4 @@
-// accessible by all users
+// TODO: add onboarding tutorial
 
 export default async function Help() {
   return <>help</>;
