+++ src/lib/utils/get-instance-path.ts
@@ -1,0 +1,6 @@
+import { instanceParams } from "@/lib/validations/params";
+
+export function getInstancePath(instanceParams: instanceParams) {
+  const { group, subGroup, instance } = instanceParams;
+  return `/${group}/${subGroup}/${instance}`;
+}
