+++ src/lib/utils/slugify.ts
@@ -1,0 +1,3 @@
+export function slugify(label: string): string {
+  return label.replaceAll("&", "and").replaceAll(" ", "-").toLowerCase();
+}
