+++ src/lib/validations/algorithm.ts
@@ -1,34 +1,7 @@
 import { AlgorithmFlag } from "@prisma/client";
 import { z } from "zod";
 
-const studentMatchingDataSchema = z.object({
-  id: z.string(),
-  preferences: z.array(z.string()),
-});
-
-const projectMatchingDataSchema = z.object({
-  id: z.string(),
-  lowerBound: z.number(),
-  upperBound: z.number(),
-  supervisorId: z.string(),
-});
-
-const supervisorMatchingDataSchema = z.object({
-  id: z.string(),
-  lowerBound: z.number(),
-  target: z.number(),
-  upperBound: z.number(),
-});
-
-export const matchingDataSchema = z.object({
-  students: z.array(studentMatchingDataSchema),
-  projects: z.array(projectMatchingDataSchema),
-  supervisors: z.array(supervisorMatchingDataSchema),
-});
-
-export const mathcingDataWithArgsSchema = matchingDataSchema.extend({
-  args: z.array(z.string()),
-});
+export const algorithmFlagSchema = z.nativeEnum(AlgorithmFlag);
 
 export const builtInAlgSchema = z.enum([
   "generous",
@@ -36,12 +9,6 @@ export const builtInAlgSchema = z.enum([
   "minimum-cost",
   "greedy-generous",
 ]);
-
-export const algorithmFlagSchema = z.nativeEnum(AlgorithmFlag);
-
-export type MatchingData = z.infer<typeof matchingDataSchema>;
-
-export type MatchingDataWithArgs = z.infer<typeof mathcingDataWithArgsSchema>;
 
 export type BuiltInAlg = z.infer<typeof builtInAlgSchema>;
 
@@ -55,36 +22,3 @@ export const algorithmSchema = z.object({
 });
 
 export type Algorithm = z.infer<typeof algorithmSchema>;
-
-export const matchingDetailsSchema = z.object({
-  student_id: z.string(),
-  project_id: z.string(),
-  project_capacities: z.object({
-    lower_bound: z.number().int(),
-    upper_bound: z.number().int(),
-  }),
-  preference_rank: z.number().int(),
-  supervisor_id: z.string(),
-  supervisor_capacities: z.object({
-    lower_bound: z.number().int(),
-    target: z.number().int(),
-    upper_bound: z.number().int(),
-  }),
-});
-
-export type MatchingDetails = z.infer<typeof matchingDetailsSchema>;
-
-export const serverResponseSchema = z.object({
-  profile: z.array(z.number()),
-  degree: z.number(),
-  size: z.number(),
-  weight: z.number(),
-  cost: z.number(),
-  costSq: z.number(),
-  maxLecAbsDiff: z.number(),
-  sumLecAbsDiff: z.number(),
-  matching: z.array(matchingDetailsSchema),
-  ranks: z.array(z.number()),
-});
-
-export type ServerResponse = z.infer<typeof serverResponseSchema>;
